requirement.txt
Cheetah==2.4.4
Django==1.6.5
Landscape-Client==14.01
Markdown==2.4.1
PAM==0.4.2
PyYAML==3.10
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
django-filter==0.7
django-rest-framework-mongoengine==1.3.2
djangorestframework==2.3.14
djangorestframework-gis==0.3
html5lib==0.999
jsonpatch==1.3
jsonpointer==1.0
mongoengine==0.8.7
oauth==1.0.1
prettytable==0.7.2
pyOpenSSL==0.13
pycurl==7.19.3
pymongo==2.7.2
pyproj==1.9.3
pyserial==2.6
python-apt==0.9.3.5
python-debian==0.1.21-nmu2ubuntu2
requests==2.2.1
six==1.5.2
ssh-import-id==3.21
suds-jurko==0.6
urllib3==1.7.1
virtualenv==1.11.4
wsgiref==0.1.2
zope.interface==4.0.5
